Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
pytz
flask_jwt_extended
flask_cors
psycopg2